+++ NEWS.md
@@ -8,6 +8,7 @@ Cite all releases with doi [10.5281/zenodo.5475443](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
+* Better matrix equality tests that are now agnostic to matrix class.
 * Improved speed of `collapse_to_matrices()` by a factor of 3.